+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.25.1 (Unreleased)
 
 BUG FIXES:
+* r/workspace: Fix remote state consumer regression for Terraform Enterprise ([#303](https://github.com/hashicorp/terraform-provider-tfe/pull/303))
 * r/organization: Ignore diffs in name case sensitivity ([#300](https://github.com/hashicorp/terraform-provider-tfe/pull/300))
 
 ## 0.25.0 (April 29, 2021)